Dummy Sign-Up Form
I created a sign-up form for The Odin Project.

This project allowed me to practice form creation and formatting. I also incorporated some additional CSS styling techniques learnt in the previous lessons to create this sign-up page.

I included some input validation in the form. Fields turn red when the parameters are not met and green when they are.